requirements.txt
streamlit
faiss-cpu
google-genai
python-docx
python-pptx
PyPDF2
python-dotenv
sentence-transformers
langchain-text-splitters
numpy
openai